Refactor the 404 and index layouts
