+++ app/routes/$.tsx
@@ -1,1 +1,5 @@
-export { default, loader } from './404.js';
+import { redirect } from '@remix-run/node';
+
+export async function loader() {
+  throw redirect('/404');
+}
